Move validator tests into their own explicit file (#11590)
[upstream:94d7abdcfa1c272f1e62d3859edca54e181959aa]
